refactor: consolidate FastExecutionRequest into ArbOpportunity

+++ controller/src/confirm_log.rs
@@ -40,7 +40,7 @@ pub struct ConfirmationRecord {
     pub yes_price_cents: u16,
     pub no_price_cents: u16,
     pub profit_cents: i16,
-    pub max_contracts: i64,
+    pub max_contracts: u16,
     pub detection_count: u32,
     pub kalshi_url: String,
     pub poly_url: String,
@@ -93,7 +93,8 @@ impl ConfirmationLogger {
         &self.file_path
     }
 
-    /// Get the number of records logged
+    /// Get the number of records logged.
+    /// Useful for summary statistics when closing a confirmation session.
     #[allow(dead_code)]
     pub fn record_count(&self) -> usize {
         self.record_count
